Scoreboard working locally
