+++ scoreboard/app/Http/Middleware/CorsMiddleware.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Middleware;
+
+use Closure;
+
+class CorsMiddleware {
+
+    public function handle($request, Closure $next)
+    {
+        $response = $next($request);
+        $response->header('Access-Control-Allow-Origin', '*');
+
+        return $response;
+    }
+}
