--- src/scoreboard.js
@@ -1,11 +1,0 @@
-// Javascript
-import 'bootstrap/dist/js/bootstrap.min';
-
-// Stylesheets
-import 'bootstrap/dist/css/bootstrap.min.css';
-
-
-$(document).ready(function ()
-{
-    console.log('scoreboard ready!');
-});
